Provide way to switch between transient attribute and not.
Requires editing the model in memory before using it, something I haven't done before. The trick is you have to copy the model you load from the bundle!

It doesn't seem possible to detach an existing persistent store from a coordinator and attach it to a new one -- all the APIs assume you can create a *new* store object and just point it at the actual persistent data. This falls down for in-memory stores! But it's also not that big of a deal.